Moved engine.config in commonlib
