+++ openquake/commonlib/config.py
