+++ openquake/engine/engine.py
@@ -28,9 +28,9 @@ from openquake.baselib.performance import Monitor
 from openquake.risklib import valid
 from openquake.commonlib import parallel, readinput
 from openquake.commonlib.oqvalidation import OqParam
-from openquake.commonlib import export, datastore
+from openquake.commonlib import export, datastore, config
 from openquake.calculators import base, views
-from openquake.engine import logs, config
+from openquake.engine import logs
 
 TERMINATE = valid.boolean(
     config.get('celery', 'terminate_workers_on_revoke') or 'false')
